Create a new ShinyQDA application

Create a new ShinyQDA application

Usage

new_app(
  name = "My_ShinyQDA",
  dir = ifelse(interactive(), rstudioapi::selectDirectory(caption =
    "Select location for ShinyQDA application"), getwd()),
  qda_data,
  id_column = find_primary_key_column(qda_data),
  text_column = find_document_column(qda_data),
  initialize_sentiment_dictionaries = TRUE,
  run_app = interactive(),
  ...
)

Arguments

name

name of the application directory.

dir

directory where the application will be created.

qda_data

a data.frame with the text data.

id_column

the name of the column in qda_data for the primary key.

text_column

the name of the column in qda_data containing the text data.

initialize_sentiment_dictionaries

whether to copy the sentiment dictionaries into the app directory.

run_app

if TRUE the Shiny application will be started once it has been created.

...

other parameters passed to qda().

Value

Returns a qda() object.