Data object for qualitative data analysis.

This function creates an object to manipulate data for qualitative analysis. The data is stored in a SQLite database by default, but can be configures to work with any SQL database through the DBI R package interface. Although you can manipulate the SQL database directly, we recommend using the methods (functions) defined in the returned object. This will ensure data integrity.

Usage

qda(
  file,
  users_passphrase = "ShinyQDA",
  users = c("admin"),
  users_names = users,
  users_passwords = rep("password", length(users)),
  users_is_admin = rep(TRUE, length(users))
)

Arguments

file

the filename and path to the file this function will save data to.

users_passphrase

passpharse used to encrypt the user authentication table. See shinymanager::create_db() for more info.

users

a vector of usernames who can login to the ShinyQDA app.

users_names

a vector of names for the users.

users_passwords

a vector of passwords for the users.

users_is_admin

a logicial vector indicating if the user has administrator privileges.

Value

a list with functions to access and edit qualitiative data.

Methods

log(coder, table, description, timestamp)

get_log()

get_last_update()

add_text(df, id_col, text_col, overwrite)

get_text(id)

add_coding(id, text, start, end, codes, coder)

update_coding(coding_id, codes)

delete_coding(coding_id)

get_codings(id, coding_id)

get_codes()

add_codes(codes, colors, descriptions)

update_code(code, color, description, parent)

add_code_question(stem, type, order, options)

delete_code_question(stem)

get_code_questions()

get_code_question_responses(coding_id, id)

delete_code_question_responses(coding_id)

add_code_question_response(coding_id, stem, answer, coder)

update_code_question_response(coding_id, stem, answer, coder)

add_text_question(stem, type, order, options)

get_text_questions()

delete_text_question(stem)

get_text_question_responses(id, coder)

delete_text_question_responses(id, coder)

update_text_question_response(id, stem, new_answer, coder)

add_text_question_response(id, stem, answer, coder)

get_rubrics()

get_rubric(rubric_name)

add_rubric(rubric_name, description, rubric, scoring_levels, enabled)

update_rubric_criteria(rubric_name, criteria, score_level, description)

delete_rubric_response(rubric_name, qda_id, coder, criteria)

add_rubric_response(rubric_name, qda_id, coder, criteria, score)

get_rubric_responses(rubric_name, qda_id, coder)

get_assignments(coder, id)

get_coders()